add abstract PublicRoute and use for others
