+++ core/domain/entities/routing/handlers/frontend/FrontendRequests.php
@@ -3,10 +3,6 @@
 namespace EventEspresso\core\domain\entities\routing\handlers\frontend;
 
 use EE_Dependency_Map;
-use EE_Maintenance_Mode;
-use EventEspresso\core\domain\entities\routing\handlers\Route;
-use EventEspresso\core\services\loaders\LoaderInterface;
-use EventEspresso\core\services\request\RequestInterface;
 
 /**
  * Class FrontendRequests
@@ -14,15 +10,10 @@ use EventEspresso\core\services\request\RequestInterface;
  *
  * @package EventEspresso\core\domain\entities\routing\handlers\frontend
  * @author  Brent Christensen
- * @since   \$VID:$
+ * @since   $VID:$
  */
-class FrontendRequests extends Route
+class FrontendRequests extends PublicRoute
 {
-
-    /**
-     * @var EE_Maintenance_Mode $maintenance_mode
-     */
-    private $maintenance_mode;
 
 
     /**
@@ -36,25 +27,6 @@ class FrontendRequests extends Route
     public function matchesCurrentRequest()
     {
         return ($this->request->isFrontend() || $this->request->isFrontAjax()) && ! $this->maintenance_mode->level();
-    }
-
-
-    /**
-     * FrontendRequests constructor.
-     *
-     * @param EE_Dependency_Map   $dependency_map
-     * @param EE_Maintenance_Mode $maintenance_mode
-     * @param LoaderInterface     $loader
-     * @param RequestInterface    $request
-     */
-    public function __construct(
-        EE_Dependency_Map $dependency_map,
-        EE_Maintenance_Mode $maintenance_mode,
-        LoaderInterface $loader,
-        RequestInterface $request
-    ) {
-        $this->maintenance_mode = $maintenance_mode;
-        parent::__construct($dependency_map, $loader, $request);
     }
